+++ release_notes.md
@@ -4,7 +4,7 @@
 2. Add `tx_hash`, `user_id`, `query_id` fields to `/v1/withdrawal/status{?id}` method response
 3. Add `failed` status to `/v1/withdrawal/status{?id}` method and withdrawal processor do not retry failed withdrawals (retry only for lost external messages)
 4. New method `/v1/deposit/income{?tx_hash}` to find income by `tx_hash`
-5. New method `/v1/balance{?currency,address}` to get account balance from blockchain and pending and processing amounts for hot wallet
+5. New method `/v1/balance{?currency,address}` to get account balance from blockchain (and account status for TON) and pending and processing amounts for hot wallet
 6. Push notifications (webhook or rabbitmq) after save to db (see new notification logic in `README.md` and `technical_notes.md`)
 7. Postgres docker volume permanent by default
 8. Audit log error counters with prometheus metrics